% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best-fit-glm.R
\name{bestFormulaGlm}
\alias{bestFormulaGlm}
\title{Generate best formula via glm}
\usage{
bestFormulaGlm()
}
\value{
The best model
}
\description{
Assumes delay.data exists
}
\examples{
bestFormulaGlm()
}
